Simplify `as unknown as ...` cast in db/index.ts (#307)
Switch to a slightly more "honest" cast of `globalThis` in the db index file.

`globalThis` _is_ an object which has an optional `prisma` property - it doesn't make much difference in this case but casting to `as unknown` allows making a slightly more unsafe cast, which isn't needed here - at least I don't think it is.